requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
Flask==3.1.0
frozenlist==1.6.0
future==1.0.0
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
haversine==2.9.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
line-bot-sdk==3.17.1
MarkupSafe==3.0.2
multidict==6.4.3
numpy==2.0.2
openai==1.79.0
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
selenium==4.32.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.21.0
